requirements.txt
# Core runtime dependencies
# Note: install a CUDA-specific torch build separately if needed.
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
pyyaml>=6.0.0
